+++ lib/src/core/photo_view_image_core.dart
@@ -1,11 +1,10 @@
 import 'package:flutter/widgets.dart';
-import 'package:photo_view/photo_view.dart';
-import 'package:photo_view/src/controller/photo_view_controller.dart';
-import 'package:photo_view/src/controller/photo_view_controller_delegate.dart';
-import 'package:photo_view/src/utils/photo_view_hero_attributes.dart';
-import 'package:photo_view/src/utils/photo_view_utils.dart';
 
-import 'photo_view_gesture_detector.dart';
+import 'package:photo_view/photo_view.dart' show PhotoViewControllerBase, PhotoViewControllerValue, PhotoViewHeroAttributes, PhotoViewScaleStateController, ScaleStateCycle;
+import 'package:photo_view/src/controller/photo_view_controller_delegate.dart';
+import 'package:photo_view/src/utils/photo_view_utils.dart';
+import 'package:photo_view/src/core/photo_view_gesture_detector.dart';
+import 'package:photo_view/src/core/photo_view_hit_corners.dart';
 
 typedef PhotoViewImageTapUpCallback = Function(
   BuildContext context,
@@ -83,7 +82,7 @@ class PhotoViewCore extends StatefulWidget {
 }
 
 class PhotoViewCoreState extends State<PhotoViewCore>
-    with TickerProviderStateMixin, PhotoViewControllerDelegate {
+    with TickerProviderStateMixin, PhotoViewControllerDelegate, HitCornersDetector {
   Offset _normalizedPosition;
   double _scaleBefore;
   double _rotationBefore;
