+++ lib/src/utils/photo_view_hero_attributes.dart
@@ -1,4 +1,4 @@
-import 'package:flutter/material.dart';
+import 'package:flutter/widgets.dart';
 
 /// Data class that holds the attributes that are going to be passed to
 /// [PhotoViewImageWrapper]'s [Hero].
